Add files via upload
Here is an example of predicted model results compared to the train data fed to the program. This is an automated output of the program meant to help check the similarity of the output. If not similar in distribution at all or exactly matching, the model may be under or over fitted.